+++ Sources/RichTextKit/_Deprecated/RichTextContext+Deprecated.swift
@@ -1,0 +1,9 @@
+#if os(iOS) || os(macOS) || os(tvOS)
+import SwiftUI
+
+public extension RichTextContext {
+
+    @available(*, deprecated, renamed: "highlightingStyle")
+    var standardHighlightingStyle: RichTextHighlightingStyle { highlightingStyle }
+}
+#endif